pyproject.toml
[tool.poetry]
name = "lesbot"
version = "0.1.0"
description = ""
authors = ["Maximilian <git@eulenbu.de>"]
packages = [{ include = "lesbot" }]
package-mode = false

[tool.poetry.scripts]
lesbot = "lesbot.app:run"

[tool.poetry.dependencies]
python = "^3.12"
discord-py = "^2.3.2"
python-dotenv = "^1.0.1"
python-magic = "^0.4.27"
pydantic = "^2.7.1"
requests = "^2.31.0"
loguru = "^0.7.2"


[tool.poetry.group.dev.dependencies]
ruff = "^0.4.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
